Deal Type system

+++ database/migrations/2023_04_26_095428_create_venue_types_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description')->nullable();
             $table->text('notes')->nullable();
-            $table->smallInteger('system')->nullable();
+            $table->integer('system')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
